Better Python3 test solution
