+++ tests/conftest.py
@@ -1,6 +1,10 @@
 import pytest
 import os
+import sys
 import tempfile
+
+if sys.version_info[0] < 3:
+    collect_ignore = ['test_3_cli.py']
 
 SDIR = os.path.dirname(os.path.abspath(__file__))
